Move error and result to separate files
