+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/AccessToken/TicketValidationAccessTokenProcessor.swift
@@ -6,12 +6,6 @@ import Foundation
 import ENASecurity
 import SwiftJWT
 
-struct TicketValidationAccessTokenResult {
-	let accessToken: String
-	let accessTokenPayload: TicketValidationAccessToken
-	let nonceBase64: String
-}
-
 struct TicketValidationAccessTokenProcessor {
 
 	// MARK: - Init
