+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/AccessToken/TicketValidationAccessTokenResult.swift
@@ -1,0 +1,13 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import ENASecurity
+import SwiftJWT
+
+struct TicketValidationAccessTokenResult {
+	let accessToken: String
+	let accessTokenPayload: TicketValidationAccessToken
+	let nonceBase64: String
+}
